feat(hvac): native 3-byte 22F1 fan payloads and L3/L7 architecture resolution
- Add native 3-byte 22F1 fan mode payload support for schemes requiring a mode_max byte.
- Replace magic string payload triggers with an explicit `legacy_format` boolean flag.
- Refactor `tests/tests_tx/test_command.py` to strictly adhere to the Arrange, Act, Assert (AAA) pattern.
- Resolve L3/L7 circular import deadlocks, exposed by isolated test collection, via PEP 562 lazy loading in `src/ramses_rf/__init__.py`.
- Add `import-linter` configuration to `pyproject.toml` to map future architectural layer boundaries.
