+++ src/ramses_cli/discovery.py
@@ -15,7 +15,8 @@ from ramses_rf import exceptions as exc
 from ramses_rf.const import SZ_SCHEDULE, SZ_ZONE_IDX
 from ramses_rf.devices import Fakeable
 from ramses_rf.protocol.opentherm import OTB_DATA_IDS
-from ramses_tx import CODES_SCHEMA, Command, DeviceIdT, Priority
+from ramses_rf.protocol_schema import CODES_SCHEMA
+from ramses_tx import Command, DeviceIdT, Priority
 from ramses_tx.typing import PayloadT
 
 from ramses_rf.const import (  # noqa: F401, isort: skip, pylint: disable=unused-import
