+++ src/ramses_rf/__init__.py
@@ -14,28 +14,32 @@ Works with (amongst others):
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
-from ramses_tx import Address, Command, Packet  # noqa: F401
-
-from . import exceptions  # noqa: F401
-from .config import GatewayConfig  # noqa: F401
-from .devices import Device  # noqa: F401
-from .exceptions import CommandInvalid  # noqa: F401
-from .gateway import Gateway  # noqa: F401
-from .messages import Message
-from .version import VERSION  # noqa: F401
-
-from .const import (  # noqa: F401, isort: skip, pylint: disable=unused-import
-    I_,
-    RP,
-    RQ,
-    W_,
-    Code,
+from . import exceptions
+from .const import I_, RP, RQ, W_, Code
+from .protocol.ramses import (
+    _2411_PARAMS_SCHEMA,
+    CODES_BY_DEV_SLUG,
+    CODES_SCHEMA,
+    SZ_DATA_TYPE,
+    SZ_DATA_UNIT,
+    SZ_DESCRIPTION,
+    SZ_MAX_VALUE,
+    SZ_MIN_VALUE,
+    SZ_PRECISION,
 )
+from .version import VERSION
 
 if TYPE_CHECKING:
-    from .const import IndexT, VerbT  # noqa: F401, pylint: disable=unused-import
+    from ramses_tx import Address, Command, Packet
+
+    from .config import GatewayConfig
+    from .const import IndexT, VerbT
+    from .devices import Device
+    from .exceptions import CommandInvalid
+    from .gateway import Gateway
+    from .messages import Message
 
 
 __all__ = [
@@ -49,6 +53,16 @@ __all__ = [
     "Device",
     "Message",
     "Packet",
+    # Schema-related constants
+    "SZ_DATA_UNIT",
+    "SZ_DESCRIPTION",
+    "SZ_DATA_TYPE",
+    "SZ_MAX_VALUE",
+    "SZ_MIN_VALUE",
+    "SZ_PRECISION",
+    "_2411_PARAMS_SCHEMA",
+    "CODES_BY_DEV_SLUG",
+    "CODES_SCHEMA",
     #
     "I_",
     "RP",
@@ -65,6 +79,48 @@ __all__ = [
 ]
 
 _LOGGER = logging.getLogger(__name__)
+
+
+def __getattr__(name: str) -> Any:
+    """Lazy-load heavy L7 domain objects and L3 primitives.
+
+    :param name: The name of the attribute to retrieve.
+    :type name: str
+    :return: The requested domain object or primitive class.
+    :rtype: Any
+    :raises AttributeError: If the requested attribute is not exported.
+    """
+    if name in ("Address", "Command", "Packet"):
+        import ramses_tx
+
+        return getattr(ramses_tx, name)
+
+    if name == "Gateway":
+        from .gateway import Gateway
+
+        return Gateway
+
+    if name == "GatewayConfig":
+        from .config import GatewayConfig
+
+        return GatewayConfig
+
+    if name == "Device":
+        from .devices import Device
+
+        return Device
+
+    if name == "Message":
+        from .messages import Message
+
+        return Message
+
+    if name == "CommandInvalid":
+        from .exceptions import CommandInvalid
+
+        return CommandInvalid
+
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
 class GracefulExit(SystemExit):
