+++ src/ramses_rf/state/store.py
@@ -28,11 +28,12 @@ from typing import TYPE_CHECKING, Any, NewType, cast
 
 import orjson
 
-from ramses_tx import CODES_SCHEMA, RP, RQ, Code, Packet
+from ramses_tx import RP, RQ, Code, Packet
 
 from ..exceptions import DatabaseQueryError
 from ..messages.base import Message
 from ..messages.core import Message as CoreMessage
+from ..protocol_schema import CODES_SCHEMA
 from ..routing import StateHeader
 from ..sqlite_worker import PacketLogEntry, SQLiteWorker
